# perf(core): use a reverse lookup in `forgetTransientResolutions`