**What's the problem this PR addresses?**

When `Project.forgetResolution` is called with a locator it loops over `Project.storedResolutions` to find all descriptors that point to that locator, this becomes expensive when called multiple times, which `Project.forgetTransientResolutions` potentially does.

**How did you fix it?**

Create a reverse mapping of `Project.storedResolutions` to efficiently lookup all descriptors pointing to a locator and delete them directly in `Project.forgetTransientResolutions`.

**Results**

Tested on the [DefinitelyTyped repo](https://github.com/jakebailey/DefinitelyTyped/tree/b09f694f323b0c5aa80c97401815f7741903be77) with its 8629 workspaces (matching only `types/*`):

```console
$ YARN_IGNORE_PATH=1 hyperfine -w 1 "node ./after.cjs --mode skip-build" "node ./before.cjs --mode skip-build"
Benchmark 1: node ./after.cjs --mode skip-build
  Time (mean ± σ):     15.802 s ±  0.437 s    [User: 12.243 s, System: 10.097 s]
  Range (min … max):   14.979 s … 16.265 s    10 runs

Benchmark 2: node ./before.cjs --mode skip-build
  Time (mean ± σ):     17.260 s ±  0.471 s    [User: 13.692 s, System: 9.957 s]
  Range (min … max):   16.425 s … 18.000 s    10 runs

Summary
  'node ./after.cjs --mode skip-build' ran
    1.09 ± 0.04 times faster than 'node ./before.cjs --mode skip-build'
```

**Checklist**
- [x] I have read the [Contributing Guide](https://yarnpkg.com/advanced/contributing).
- [x] I have set the packages that need to be released for my changes to be effective.
- [x] I will check that all automated PR checks pass before the PR gets reviewed.